Updated Handlers and reports

+++ routes/web.php
@@ -21,13 +21,18 @@ Route::get('/callback/google', [App\Http\Controllers\login::class, 'GoogleCallba
 Route::get('/callback/facebook', [App\Http\Controllers\login::class, 'FacebookCallback'])->name('facebook.callback');
 Route::get('callback/microsoft', [App\Http\Controllers\login::class, 'MicrosoftCallback'])->name('microsoft.callback');
 
-Route::get('/users', [App\Http\Controllers\Users::class, 'index'])->middleware('isAdmin')->name('users');
-Route::get('/user/config', [App\Http\Controllers\Users::class, 'config'])->middleware('isLogged')->name('users.config');
-Route::get('/user/{id}', [App\Http\Controllers\Users::class, 'show'])->name('users.show');
-Route::delete('/user/{id}', [App\Http\Controllers\Users::class, 'delete'])->name('users.delete');
-Route::post('/user', [App\Http\Controllers\Users::class, 'update'])->name('users.update');
+Route::get('/users', [App\Http\Controllers\Users::class, 'index'])->name('users')->middleware('isAdmin');
+Route::get('/user/config', [App\Http\Controllers\Users::class, 'config'])->name('users.config')->middleware('isVolunteer');
+Route::get('/user/{id}', [App\Http\Controllers\Users::class, 'show'])->name('users.show')->middleware('isAdmin');
+Route::delete('/user/{id}', [App\Http\Controllers\Users::class, 'delete'])->name('users.delete')->middleware('isAdmin');
+Route::post('/user', [App\Http\Controllers\Users::class, 'update'])->name('users.update')->middleware('isAdmin');
 
+Route::get('/reports', [App\Http\Controllers\Reports::class, 'index'])->name('reports')->middleware('isAdmin');
+Route::get('/reports/{userid}', [App\Http\Controllers\Reports::class, 'listUser'])->name('reports.userList')->middleware('isAdminorVolunteer');
+Route::get('/reports/{userid}/{reportid}', [App\Http\Controllers\Reports::class, 'show'])->name('reports.show')->middleware('isAdminorVolunteer');
+Route::post('/weeklyreport/update', [App\Http\Controllers\Reports::class, 'updateweekly'])->name('reports.updateweekly')->middleware('isVolunteer');
 
-Route::get('/reports', [App\Http\Controllers\Reports::class, 'index'])->name('reports');
-Route::get('/reports/{userid}', [App\Http\Controllers\Reports::class, 'listUser'])->name('reports.userList');
-Route::get('/reports/{userid}/{reportid}', [App\Http\Controllers\Reports::class, 'show'])->name('reports.show');
+//Temporal home routing
+Route::get('/home', function () {
+    return view('home');
+})->name('home')->middleware('auth');
